Set explicit test dependency on Nokogiri 1.6.x
Our current stack in Go Pipelines uses a very old of LibXML2, this
causes issues when using Nokogiri over 1.6 (1.7 or 1.8) as when parsing
websites removes the expected return carriages and causes assertions
failing.

There are plans to upgrade Go Pipelines environments so a more modern
LibXML2 version is used and we can upgrade to Nokogiri newer versions,
but until that point, we actually depend on using 1.6 version.

This commit raises/sets that explicit dependency.